[Refactor]: userchallenge adapter 분리
