+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/infrastructure/userchallenge/UserChallengeRepository.java
@@ -1,4 +1,4 @@
-package org.haedal.zzansuni.infrastructure.userchallenge.userchallenge;
+package org.haedal.zzansuni.infrastructure.userchallenge;
 
 import java.time.LocalDate;
