+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/infrastructure/userchallenge/adapter/UserChallengeStoreImpl.java
@@ -1,8 +1,9 @@
-package org.haedal.zzansuni.infrastructure.userchallenge.userchallenge;
+package org.haedal.zzansuni.infrastructure.userchallenge.adapter;
 
 import lombok.RequiredArgsConstructor;
 import org.haedal.zzansuni.domain.userchallenge.UserChallenge;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeStore;
+import org.haedal.zzansuni.infrastructure.userchallenge.UserChallengeRepository;
 import org.springframework.stereotype.Component;
 
 @Component
